Mistake in the source string
Also I feel the "Fly mode is enabled as part of the creative mode to allow for easier building __and servers__ allow the use of world editing plugins..." is a bit wired. You can change it to "Fly mode is enabled as part of the creative mode to allow for easier building. _Servers also_ allow the use of world editing plugins which makes the building process even easier." if it seems better...